Public fifo only opens once. Timeout on opening fifo.

+++ src/include/parse.h
@@ -1,6 +1,8 @@
 #ifndef PROJECT_INCLUDE_PARSE_H_
 #define PROJECT_INCLUDE_PARSE_H_
 
-int parseCommand(int argc, char *argv[], char* fifoname[], int *nsecs);
+#include <sys/types.h>
+
+int parseCommand(int argc, char *argv[], char *fifoname[], size_t *nsecs);
 
 #endif
