+++ src/parse.c
@@ -7,7 +7,7 @@
 #include <sys/types.h> // mode_t
 #include <unistd.h> // getopt(), optind
 
-int parseCommand(int argc, char *argv[], char* fifoname[], int *nsecs) {
+int parseCommand(int argc, char *argv[], char* fifoname[], size_t *nsecs) {
 
     if (argv == NULL || fifoname == NULL || nsecs == NULL) return -1;
